refactor: 테스트 코드 패키지 변경, 프로필페이지 이메일 정보 노출
-
