+++ src/main/resources/templates/user/profile.html
@@ -86,7 +86,7 @@
                         <div class="media-body">
                             <h4 class="media-heading">{{userId}}</h4>
                             <p>
-                                <a href="#" class="btn btn-xs btn-default"><span class="glyphicon glyphicon-envelope"></span>&nbsp;{{userId}}</a>
+                                <a href="#" class="btn btn-xs btn-default"><span class="glyphicon glyphicon-envelope"></span>&nbsp;{{email}}</a>
                             </p>
                         </div>
                     </div>
